pyproject.toml
[tool.poetry]
name = "financial-assistant-llm-agent"
version = "0.1.0"
description = "A conversational finance Q&A project powered by language models."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"
readme = "README.md"
# repository = "https://github.com/yourusername/convfinqa-llm-project"
# documentation = "https://yourdocs.example.com"

[tool.poetry.dependencies]
python = "^3.10.16"

# Language model interfaces and tools
langchain = "0.3.23"
langchain-core = "0.3.51"
langchain-openai = "0.3.12"
langchain-anthropic = "0.3.10"
langchain-google-vertexai = "2.0.19"
langchain-text-splitters = "0.3.8"
langsmith = "0.3.28"

# Cloud + LLM libraries
openai = "1.72.0"
anthropic = "0.49.0"
google-cloud-aiplatform = "1.88.0"
google-cloud-storage = "2.19.0"
google-cloud-bigquery = "3.31.0"

# General libraries
python-dotenv = "1.1.0"
requests = "2.32.3"
pandas = "2.2.3"
numpy = "2.2.4"
tiktoken = "0.9.0"
shapely = "2.1.0"
validators = "0.34.0"
orjson = "3.10.16"

# Hugging Face + Tokenizers (optional)
huggingface-hub = "0.30.2"
tokenizers = "0.21.1"

# Required support libs
pydantic = "2.11.3"
pydantic_core = "2.33.1"
httpx = "0.28.1"
httpx-sse = "0.4.0"
tenacity = "9.1.2"
typing-extensions = "4.13.2"
typing-inspection = "0.4.0"

# Development and testing
[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
pytest-mock = "^3.10"
black = "^23.0"
flake8 = "^5.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
